move composable to store & fix bug
